feat: DbtProjects can now create DbtProject subprojects
This is the second step in recursive projects. This involces having create_subproject return a new DbtProject. There are a couple of major things that need to happen next, however. 1) The subproject returned should be a different type so we know that it cannot leverage a dbt runner. 2) The subproject can be materialized into a real project.