refactor: Get rid of DBTFileManager in favor of YAML and Raw-specific file managers
Why get rid of the DbtFileManager? It returned mutually exclusive types! When this
happens, we're stuck doing type checks every time we read a file, which is a nightmare.
Secondly, the DbtFileManager had assumptions about realtive paths being used in oeprations,
which as of ChangeSets is NO LONGER THE CASE. Now, absolute paths are used everywhere. As a
result, it made sense to split out our concerns quite a bit such that there is a
RawFileManager that handles I/O of raw data, and a YamlFileManager that adds in
a YAML parsing layer to keep our code clean.